correctly selecting nodes
