+++ application/features/documents_graph/src/commonMain/kotlin/io/writeopia/documents/graph/navigation/NavigateDocumentGraph.kt
@@ -3,7 +3,6 @@ package io.writeopia.documents.graph.navigation
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -11,7 +10,6 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import io.writeopia.common.utils.Destinations
 import io.writeopia.documents.graph.di.DocumentsGraphInjection
-import io.writeopia.documents.graph.extensions.toGraph
 import io.writeopia.forcegraph.ForceDirectedGraph
 
 fun NavController.navigateToForceGraph() {
@@ -30,16 +28,10 @@ fun NavGraphBuilder.documentsGraphNavigation(
         val state by viewModel.graphSelectedState.collectAsState()
 
         BoxWithConstraints(modifier = Modifier.fillMaxSize()) {
-            val graph by derivedStateOf {
-                state.toGraph(
-                    this.maxWidth.value,
-                    this.maxHeight.value
-                )
-            }
 
             ForceDirectedGraph(
-                nodes = graph.nodes,
-                links = graph.links,
+                nodes = state.nodes,
+                links = state.links,
                 onNodeSelected = viewModel::selectNode
             )
         }
